+++ tests/Utils/Middleware/ExceptionMiddleware.php
@@ -3,11 +3,9 @@ namespace WoohooLabsTest\Harmony\Utils\Middleware;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use WoohooLabs\Harmony\Harmony;
-use WoohooLabs\Harmony\Middleware\MiddlewareInterface;
 use WoohooLabsTest\Harmony\Utils\Exception\TestException;
 
-class ExceptionMiddleware implements MiddlewareInterface
+class ExceptionMiddleware
 {
     protected $text;
 
@@ -30,11 +28,11 @@ class ExceptionMiddleware implements MiddlewareInterface
     /**
      * @param \Psr\Http\Message\ServerRequestInterface $request
      * @param \Psr\Http\Message\ResponseInterface $response
-     * @param \WoohooLabs\Harmony\Harmony $next
+     * @param callable $next
      * @return \Psr\Http\Message\ResponseInterface
      * @throws \WoohooLabsTest\Harmony\Utils\Exception\TestException
      */
-    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, Harmony $next)
+    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next)
     {
         throw new TestException($this->text);
     }