+++ tests/Utils/Middleware/FakeMiddleware.php
@@ -3,10 +3,8 @@ namespace WoohooLabsTest\Harmony\Utils\Middleware;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use WoohooLabs\Harmony\Harmony;
-use WoohooLabs\Harmony\Middleware\MiddlewareInterface;
 
-class FakeMiddleware implements MiddlewareInterface
+class FakeMiddleware
 {
     protected $text;
 
@@ -29,10 +27,10 @@ class FakeMiddleware implements MiddlewareInterface
     /**
      * @param \Psr\Http\Message\ServerRequestInterface $request
      * @param \Psr\Http\Message\ResponseInterface $response
-     * @param \WoohooLabs\Harmony\Harmony $next
+     * @param callable $next
      * @return \Psr\Http\Message\ResponseInterface
      */
-    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, Harmony $next)
+    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, callable $next)
     {
         return $next();
     }